Cleaning dashboardcontroller (not done yet)
